Updates to add boot, loader/updater, and app.
Updated memory map to have three sections, boot, loader, and app roms.

Added image lib so that each image (ex. app) can have an image header containing some information for validating the image.

Boot validates the loader and jumps to it. If not valid, boot valids the next rom slot which is app, and jumps to it.

Added a test app directory for testing the bootloader.